refactor: lint issue fixed
