+++ frontend/src/App.css
@@ -1,6 +1,7 @@
 .App {
   text-align: center;
   position: relative;
+  max-height: 900px;
 }
 
 .title {
