+++ frontend/src/components/Carousel/carousel.css
@@ -22,6 +22,7 @@
   left: 0;
   bottom: 0;
   width: 100%;
+  height: 100%;
 }
 
 .carousel-item .space {
