+++ frontend/src/components/Progress/progress.tsx
@@ -53,7 +53,7 @@ const Progress: React.FC<ProgressProps> = (props: ProgressProps) => {
     <>
       <div className="progress-container">
         <div
-          role="progress-percent"
+          data-testid="progress-percent"
           className="progress-percent"
           style={{
             width,
